+++ .internal/assemblyscript/@gl/api/w2h/host.ts
@@ -1,10 +1,10 @@
 import * as char from "./char";
-import * as choices from "./choices";
 import * as controls from "./controls";
 import * as debug from "./debug";
 import * as filters from "./filters";
 import * as lights from "./lights";
 import * as map from "./map";
+import * as markers from "./markers";
 import * as npc from "./npc";
 import * as particles from "./particles";
 import * as physics from "./physics";
@@ -22,12 +22,12 @@ import * as ui from "./ui";
 
 export {
   char,
-  choices,
   controls,
   debug,
   filters,
   lights,
   map,
+  markers,
   npc,
   particles,
   physics,