+++ CHANGELOG.md
@@ -1,10 +1,10 @@
-##
+## 0.11.0 - 5/16/25
 
 - Fix AS syntax highlighting
 - Fix hotreloading level assets
-- Support async function calls in Assemblyscript
-- Async support in Assemblyscript!!!
+- Support async function calls in Assemblyscript!!
 - Bugfix with Twine Sugarcube not transpiling parenthesis
+- Marker api for checking a player's past choices
 
 ## 0.10.0 - 5/5/25